Use async thread for performPoll in NetworkStats
The performPoll will be executed in BroadcastReceiver
process. But performPoll may sync the time by NTP.
If the NTP server is not available to access this process
will suspend for 20s (depends on the NTP timeout config).
In this 20s all of the other ordered broadcast will be
delay.

This patch use async thread for performPoll. It will
return the receiver process and won't suspend other
broadcast.

Change-Id: I9605911a9994397c2db367b02a1c7aeeb03e601f
Signed-off-by: guoyin.chen <guoyin.chen@freescale.com>
Signed-off-by: zhang sanshan <b51434@freescale.com>
Signed-off-by: Wang Haoran <Haoran.Wang@freescale.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>